ai/pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = "ai flask server"
authors = ["theohong <theohong0234@gmil.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <3.11"
tensorflow = "^2.9.1"
Flask = "^2.1.2"
gevent = "^21.12.0"
requests = "^2.28.0"
Pillow = "^9.1.1"
Flask-Cors = "^3.0.10"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
